#416 move sync dataset up to brain observatory subpackage
